Refactor to separate sidebar and pages toggle and sidebar components
